refactor(rest): extract admin worker/job routes into routes/admin_workers.py
Moves the four admin worker-fleet + job-control routes (GET workers,
POST workers/prune, GET worker-packages/{image_tag:path}, POST
jobs/{job_id}/cancel) out of create_app's closure into
routes/admin_workers.py, next slice of the admin router split.

No new RestContext/deps.py surface needed — the routes read the job
queue via RestContext.queue (worker list/prune) or directly off
request.app.state (job cancel), same as inside the closure.
